Command status timestamps are in ISO-8601 format

+++ opal-gwt-client/src/main/java/org/obiba/opal/web/gwt/datetime/client/Moment.java
@@ -124,7 +124,7 @@ public class Moment extends JavaScriptObject {
    * Get the Unix EPOCH time in milliseconds.
    * @return
    */
-  public final native int valueOf() /*-{
+  public final native double valueOf() /*-{
       return this.valueOf();
   }-*/;
